requirements.txt
pandas>=2.0
requests>=2.31
python-dotenv>=1.0
streamlit>=1.37
psycopg2-binary>=2.9
SQLAlchemy>=2.0
matplotlib>=3.8
jupyter>=1.0
